Problem statement overview
Our case study is the Turkana community living in Northern Kenya near Lake Turkana. The area is highly vulnerable to floods due to excessive rainfall predicted in East Africa as a result of climate change according to UNFCCC reports. Turkana community is recognized as one of indigenous and vulnerable people by UNESCO due to their limited access to modern technology. They are living in a traditional lifestyle and overdependence on nature for livelihood. Recently, Lake Turkana got flooded and displaced more than 1000 people further increasing the vulnerability of this local community. Any slight trigger in the environment, like climate change the Turkana community are directly affected.

The users of this technology are Turkana community who live in remote place with limited communication in Kenya. This application is most appropriate will assist in improving awareness about flood situation in this Turkana area to help the people be well prepared in advance to reduce causalities and loss of assets. The flood disaster management App will also assist in informing Government, NGOs and private sector to intervene and solicit the help Turkana community need from well-wishers. This will enable them recover quickly and go on with their business as usual after floods.

The mobile service provider, the Safaricom will be the one to host this platform to ease communication. This technology can be replicated in other areas that experience floods as a result of climate change.